syzkaller/executor: pin the main thread to CPU0 and other threads to CPUs excluding CPU0
I observed a throughput issue that after the fork-ed process finishes
its job, waitpid() in the main thread takes a long time to succeed. I
tried to address the issue by pinning the main thread to CPU0 and
other threads to CPUs excluding CPU0, but it does not work as
expected. We need to solve the issue for the fuzzing throughput.

Pinning is required for QCSCHED anyway.