syzkaller/syz-fuzzer: pkg/signal: improve struct ReadFrom
Previous implementation is limited to calculate the length of ReadFrom
because of the nested map implementation; It takes O(N). For easy of
calculating the length, change its implementation by using a single
map. Its drawback is that finding out all read-from reltationships
from a single instruction takes O(N), but we don't need it yet.