syzkaller/pkg/siganl: sort Accesses before hand them to SerialAccess.Add()
acc is not sorted when serializeAccess() is called by
FromAcesses(). Although SerialAccess will sort them, it is too slow
since moving elements need to copy lots of memory objects. To take
advantage of the fast path (i.e., idx == n in Add()), we sort acc here
and then hand it to serial.Add().