syzkaller/executor: workaround of the waitpid() latency issue
Possibly because of store buffer emulation, waitpid() returns pid a
long time after the worker thread executes doexit(0) (see
event_set(exit_notify) above). This severly harms the throughput of
our fuzzer, such that the executor executes only 1~2 programs for a
second. To workaroung this issue, we use exit_notify to let the loop
thread know the worker threads are going to exit.