syzkaller/prog: fix writeEpoch()
In the Razzer's threading mechanism, the only case that we don't send
execInstrEpoch is the racing call that will be executed by thread 0.
Let's determine the given call is the above case by looing all later
calls and finding a call that will be executed in the same
epoch (i.e., the given call is a racing call).