qemu/qcsched: fix a race condition qcsched_handle_breakpoint() vs qcsched_deactivate_breakpoint()
When removing a breakpoint on another CPU, kvm_remove_breakpoint_cpu()
temporary releases the iolock. This opens a chance of race condition
between this function and a function removing a breakpoint on this
CPU, and consequently, kvm_remove_breakpoint_cpu() can return
-ENOENT. Since the only location that removes breakpoints on other
CPUs is qcsched_deacitavte_breakpoint() which falsify sched.activated,
we can check sched.activated to confirm that the error code is
actually benign.