qemu/qcsched: implmenet HCALL_RESET
If an executor thread abnormally exited, a garbage schedule still
resides in the hypervisor. Fuzzer need to reset it before executing
the next schedule.

qcsched_reset() should be executed in CPU with the index 0, and all
other worker CPUs should be executed in CPU with the index other than
0. Otherwise, qcsched_reset() and other hcalls can race causing a
deadlock.