qemu/qcsched: add watchdog-breakpoint
In this project, there is no case that a breakpoint keep being hit
consecutively so far (we don't consider cases where an instruction is
executed multiple times, such as a loop; this will be fixed in the
future). So if a breakpoint is hit multiple times in a row, something
goes wrong (e.g., race condition). This watchdog is intended to detect
it early.