syzkaller/syz-fuzzer: monitor affinity for the fuzzer goroutine and proc goroutines
This is possibly not enough; we do not lock goroutines to OS threads
nor limit the number of OS threads. Nonetheless, we start the fuzzer
using the taskset util which sets the affinity for the child (i.e.,
syz-fuzzer). Since the affinity is inherited, I hope this is
enough (i.e., OS threads for syz-fuzzer inherits the affinity). Let's
monitor what will happen during the execution