syzkaller/pkg/primitive: reflect a timestamp into Communication hash
The purpose of Segment.Hash() was not to reflect timestamps of
Accesses. But this caused to let knotter.dupedComm() eliminate a
necessary Communication, and a test fail consequently (i.e.,
TestExcavateKnotsTwoSeqs with cve-2019-6974-2-seq[12]).

Reflecting timestamps into Hash() can fix the above issue. Although
this breaks the original purpose, fortunately, we use
Communication.Hash() only in knotter.dupedComm(), which makes
everything fine (at least for now). Let's do it.