qemu/qcsched: fix the assertion violation when closing a window
When we are closing the window, we don't care window manipulation
operations. In fact, without qcsched_window_{sync,
cleanup_leaft_schedpoint} the manipulation operations does not work
which causes the assertion violation in __do_shrink_entry().

To fix the assertion violation, we directly handle the entry instead
of relying on the window manipulation.