syzkaller/executor: suppress the "wrong response packet" syzfail
The syzfail seems like a race condition between hci_send_frame() and
__vhci_create_device() in a kernel. I can reproduce this syzfail with
tests/vhci_init_test.c (which is not affected by my modification) in
the lastest kernel (v5.18, 4b0986a36).  I'm not sure this is a bug or
not, but it is not our fault anyway, so let's ignore it.