qemu/qcsched: update the guest debug state after closing the window
I'm not sure whether this is required. I observe infinite failing of
removing a breakpoint (see below). My guess is in QEMU, the breakpoint
data is removed but KVM does not know about it. Updating the guest
debug state is with a hope of fixing it. Let's see what will happen.

---------->8---------
[CPU #2-1063/214300/8] Failed to remove a breakpoint (error=-2). Abort a schedule
[CPU #2-1063/214300/8] Closing the window
[CPU #2-1063/214301/8] Failed to remove a breakpoint (error=-2). Abort a schedule
[CPU #2-1063/214301/8] Closing the window
[CPU #2-1063/214302/8] Failed to remove a breakpoint (error=-2). Abort a schedule
[CPU #2-1063/214302/8] Closing the window
[CPU #2-1063/214303/8] Failed to remove a breakpoint (error=-2). Abort a schedule
[CPU #2-1063/214303/8] Closing the window