exp/report: add a report of INFO: task hung in usb_get_descriptor
XXX: We need to investigate this crash more. There is a same crash
title found by Syzkaller[1], and its fix is already proposed[2]. It
seems that the patch is merged into the Linus' branch at some point,
and I can see the commit in my branch (Linux mainline commit:
92f20ff72066d, my Linux branch commit: 0f3b0829949). But still we have
the same crash observed.

This may indicate an incomplete fix. We need to investivate.

[1]: https://syzkaller.appspot.com/bug?id=5015e2627c0565b479938624b4dc218ec4aac7db
[2]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=363eaa3a450abb4e63bd6e3ad79d1f7a0f717814