qemu/qcsched: workaround the rcu stall in corrupted
Possibly due to the broken logic of __handle_breakpoint_hook(), a
breakpoint keeps installed/removed at the hook (i.e.,
"__handle_breakpoint_hook" and "Reinstalling a breakpoint" keeps
printing alternately). Ideally we need to fix the logic, but as a
workaround, watchdog monitors also a breakpoint on the hook. If
watchdog fails, the breakpoint handler closes the window just returns.