qemu/qcsched: do not check qcsched_vmi_in_task() in __handle_breakpoint_hook()
If a breakpoint is hit on the hook, it must be a task context. Relax
the condition in __handle_breakpoint_hook() with a hope that
unexpected reinstall a bp happens less.

NOTE: If this works, it possibly indicates that qcsched_vmi_in_task()
is broken and vmi_lock_{acquire/release} do not work as expected.