SSBPass: handle some intrinsic functions that do not imply memory barriers
Previously, flush callbacks are instrumented before intrinsic
functions regardless of what those functions are doing. This prevents
triggering some bug such as the one in xsk.

So we need not to instrument flush callbacks in front of those
intrinsics.

TODO: My assumption is that almost all intrinsic functions do not
imply a memory barrier, and llvm.gc* intrinsics may imply a
barrier. I'm not sure this is correct.