scripts/linux/build_setup.sh: do not instrument dynamic allocas
Our FuncEntry callbacks break the entry basic block. As a result, some
static allocas turns out dynamic allocas which causes the KASAN
instrumentation for dynamic allocas. Not sure why, but this breaks the
correctness related to put_user_*(), specifically related to the ebx
clobber.

As it is unlikely that interesting bugs cause
stack-out-of-bounds, (perhaps) we can safely disable
-asan-instrument-dynamic-allocas.