kernels/linux: merge store history
At this point, I'm not sure that we can utilize store buffer in
fuzzing as it has not been tested enough. To prevent a regression,
kssb disables load prefetch by default and we can turn it on through
the debugfs interface.

TODO: Need to check the correctness of the timetstamp management