syzkaller/pkg/scheduler: always consider a given trace is a sequential trace
Now the tool has an assumption that a trace is always a sequential
trace. It is acknowledged that the assumption is not held sometimes,
for example, when the first executed syscall is blocked. This
assumption may lead us to miss some bugs, but I think it is unlikely
in the experiment.